requirements.txt
google-generativeai
python-dotenv
jinja2
markdown
